requirements.txt
absl-py==2.4.0
attrs==26.1.0
cffi==2.0.0
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cuda-toolkit==12.8.1
cycler==0.12.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.21
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.19.0.56
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.21.0
onnx-ir==0.2.1
onnxruntime-gpu==1.25.1
onnxscript==0.7.0
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==26.1
pandas==3.0.2
pillow==12.2.0
protobuf==4.25.9
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
setuptools==70.2.0
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
tensorrt_cu12==10.16.1.11
tensorrt_cu12_bindings==10.16.1.11
tensorrt_cu12_libs==10.16.1.11
tensorrt_cu13==10.16.1.11
tensorrt_cu13_bindings==10.16.1.11
tensorrt_cu13_libs==10.16.1.11
threadpoolctl==3.6.0
torch==2.11.0+cu128
torchvision==0.26.0+cu128
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0

# For the JETSON Only : 
numpy<2.0.0
scipy
pandas
mediapipe==0.10.18
tqdm
onnx
onnxscript

# 1. Create the environment with system site packages enabled
python3 -m venv .venv --system-site-packages

# 2. Activate it
source .venv/bin/activate

# pip install -r requirements_JETSON.txt

# Install Jetson-specific PyTorch dependencies
pip install setuptools Cython wheel

# Download and install the Nvidia Jetson PyTorch wheel
wget https://developer.download.nvidia.com/compute/redist/jp/v62/pytorch/torch-2.5.0a0+872d972e41.nv24.08-cp310-cp310-linux_aarch64.whl
pip install torch-2.5.0a0+872d972e41.nv24.08-cp310-cp310-linux_aarch64.whl

# Install the Nvidia Jetson ONNX Runtime wheel
wget https://developer.download.nvidia.com/compute/redist/jp/v62/onnxruntime/onnxruntime_gpu-1.19.2-cp310-cp310-linux_aarch64.whl
pip install onnxruntime_gpu-1.19.2-cp310-cp310-linux_aarch64.whl